Refactor: Extract superclass from IterableCodec and MapCodec
